\name{Arrows3D}
\alias{Arrows3D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function
}
\description{
   Internal function.
}
\usage{
Arrows3D(a, b, colo = "blue", headLength = 0.3, headWidth = 0.5, width = 1, transparent = FALSE, plano = "XY")
}

\author{Juan Carlos Ruiz Cuetos \email{bilba_t@hotmail.com}, Maria Eugenia Polo Garcia \email{mepolo@unex.es}, Pablo Garcia Rodriguez \email{pablogr@unex.es}}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
