\name{Pause}
\alias{Pause}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function Pause}
\description{
Function Pause
}
\usage{
Pause()
}
%- maybe also 'usage' for other objects documented here.
\details{
Function Pause
}
\value{
No values
}
\references{No references}
\author{Ruiz-Cuetos, J.C., \email{bilba_t@hotmail.com}, Polo, M.E., \email{mepolo@unex.es}, Rodriguez, P.G. \email{pablogr@unex.es}}
\note{No notes}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{No seealso}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function () { 
    cat("Hit <enter> to continue...")
    readline()
    invisible()
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
