\name{ToCalculateIncr3D}
\alias{ToCalculateIncr3D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Internal function.
}
\description{
   Internal function.
}
\usage{
ToCalculateIncr3D(coord)
}
%- maybe also 'usage' for other objects documented here.

\author{Felicisimo, A.M., \email{amfeli@unex.es},
Ruiz-Cuetos, J.C., \email{bilba_t@hotmail.com}, 
Polo, M.E., \email{mepolo@unex.es}, 
Cuartero, A., \email{acuartero@unex.es}, 
Rodriguez, P.G. \email{pablogr@unex.es}}


%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
