\name{ToRadians3D}
\alias{ToRadians3D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function
}
\description{
  Internal function.
}
\usage{
ToRadians3D(grades)
}
%- maybe also 'usage' for other objects documented here.

\author{Juan Carlos Ruiz Cuetos \email{bilba_t@hotmail.com}, Maria Eugenia Polo Garcia \email{mepolo@unex.es}, Pablo Garcia Rodriguez \email{pablogr@unex.es}}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
