\name{Vectors3D}
\alias{Vectors3D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function
}
\description{
  Internal function
}
\usage{
Vectors3D(vectors)
}

\author{Juan Carlos Ruiz Cuetos \email{bilba_t@hotmail.com}, Maria Eugenia Polo Garcia \email{mepolo@unex.es}, Pablo Garcia Rodriguez \email{pablogr@unex.es}}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
