\name{XYZcoor}
\docType{data}
\alias{XYZcoor}
\title{GPS Data Set}
\description{
	Dataset obtained with a GPS, UNEX Merida.
}
\usage{XYZcoor}
\format{A matrix containing 53 observations.}
\source{UNEX Merida}
\references{
	Escuela Universitaria Politecnica, Caceres.
}
\keyword{datasets}
