\name{ToSexagesimal3D}
\alias{ToSexagesimal3D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function.
}
\description{
   Internal function.
}
\usage{
ToSexagesimal3D(radians)
}

\author{Ruiz-Cuetos, J.C., \email{bilba_t@hotmail.com}, Polo, M.E., \email{mepolo@unex.es}, Rodriguez, P.G. \email{pablogr@unex.es}}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
